Revert "Feat/favContext (#7)"
This reverts commit 8ab8e80c8fbe74dde2bfbaf18723236af2ba8923.
