+++ ecommerce/src/App.js
@@ -3,18 +3,16 @@ import Navbar from "./components/navbar/index";
 import ProductsCarousel from "./components/carrusel/carrusel";
 import footer1 from "./components/footer/footer";
 import { CartProvider } from "./hook/useCart";
-import { useState, useEffect, useContext, useRef } from "react";
+import { useState, useEffect } from "react";
 import jwt_decode from "jwt-decode";
 import { toast, ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 import Footer from "./components/footer/footer";
-import { FavoritesContext, FavoritesProvider } from "./hook/useFav";
 
 function App() {
   const [isLoggedIn, setIsLoggedIn] = useState(false);
   const [showFooter, setShowFooter] = useState(false);
-  const {getFavProduct} = useContext(FavoritesContext);
-  const favoritesProviderRef = useRef();
+
   useEffect(() => {
     setTimeout(() => {
       setShowFooter(true);
@@ -53,7 +51,6 @@ function App() {
         setIsLoggedIn(true);
         console.log("Inicio de sesión exitoso");
         console.log(data);
-        getFavProduct();
         if (data.success === true) {
           toast.success("Inicio de sesión exitoso", {
             className: "mobile-toast",
@@ -92,19 +89,17 @@ function App() {
     }, 100);
   };
   return (
-    <FavoritesProvider isLoggedIn={isLoggedIn} ref={favoritesProviderRef}>
-      <CartProvider>
-        <Navbar
-          scrollToCategory={scrollToCategory}
-          isLoggedIn={isLoggedIn}
-          handleLogin={handleLogin}
-          handleLogout={handleLogout}
-        />
-        <ProductsCarousel />
-        <ToastContainer />
-        {showFooter && <Footer />}
-      </CartProvider>
-    </FavoritesProvider>
+    <CartProvider>
+      <Navbar
+        scrollToCategory={scrollToCategory}
+        isLoggedIn={isLoggedIn}
+        handleLogin={handleLogin}
+        handleLogout={handleLogout}
+      />
+      <ProductsCarousel />
+      <ToastContainer />
+      {showFooter && <Footer />}
+    </CartProvider>
   );
 }
